Some refactoring - removes DDD folder and instead keeps DDD classes directly in application folder - moves duplicate code inside DatabaseInterface to private function
