+++ src/application/EntryFactory.h
