+++ src/application/EntryRepository.cpp
