+++ src/tests/application/AbstractDatabaseInterfaceMock.h
@@ -1,8 +1,8 @@
 #ifndef SRC_ABSTRACTDATABASEINTERFACEMOCK_H
 #define SRC_ABSTRACTDATABASEINTERFACEMOCK_H
 
-#include "../../../application/AbstractDatabaseInterface.h"
-#include <gmock/gmock.h>
+#include "../../application/AbstractDatabaseInterface.h"
+#include "../../cmake-build-debug/_deps/googletest-src/googlemock/include/gmock/gmock.h"
 
 class AbstractDatabaseInterfaceMock : public AbstractDatabaseInterface {
 public:
