+++ src/tests/test_main.cpp
@@ -1,5 +1,5 @@
-#include "./plugins/encryption/crypto_tests.cpp"
-#include "./adapters/database/databaseInterface_tests.cpp"
-#include "./application/DDD/encryptor_tests.cpp"
-#include "./application/DDD/factory_tests.cpp"
-#include "./application/DDD/repository_tests.cpp"
+#include "plugins/encryption/crypto_tests.cpp"
+#include "adapters/database/databaseInterface_tests.cpp"
+#include "application/encryptor_tests.cpp"
+#include "application/factory_tests.cpp"
+#include "application/repository_tests.cpp"
